Arreglando bugs microservicio competencias rap

+++ competencias-rap-service/src/main/java/org/unicauca/competenciasrapservice/fachadaServices/services/CompetenciaService.java
@@ -6,7 +6,11 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.unicauca.competenciasrapservice.capaAccesoADatos.models.Competencia;
+import org.unicauca.competenciasrapservice.capaAccesoADatos.models.Rap;
 import org.unicauca.competenciasrapservice.capaAccesoADatos.repositories.ICompetenciaRepositorio;
+import org.unicauca.competenciasrapservice.capaAccesoADatos.repositories.IRapRepositorio;
+import org.unicauca.competenciasrapservice.capaControladores.controladorExcepciones.excepcionesPropias.EntidadNoExisteException;
+import org.unicauca.competenciasrapservice.capaControladores.controladorExcepciones.excepcionesPropias.ReglaNegocioExcepcion;
 import org.unicauca.competenciasrapservice.fachadaServices.DTO.CompetenciaDTO;
 
 import java.util.List;
@@ -17,6 +21,8 @@ import java.util.stream.Collectors;
 public class CompetenciaService implements ICompetenciaService{
     @Autowired
     private ICompetenciaRepositorio competenciaRepositorio;
+    @Autowired
+    private IRapRepositorio rapRepositorio;
     @Autowired
     private ModelMapper modelMapper;
 
@@ -69,12 +75,25 @@ public class CompetenciaService implements ICompetenciaService{
     public boolean eliminarCompetencia(int id) {
         Optional<Competencia> optionalCompetencia = competenciaRepositorio.findById(id);
         if (optionalCompetencia.isPresent()) {
-            competenciaRepositorio.deleteById(id);
+            Competencia competencia = optionalCompetencia.get();
+
+            // Verificar si existen RAPs asociados
+            System.out.println(competencia.getId());
+            List<Rap> rapsAsociados = rapRepositorio.findByCompetenciaId(competencia.getId());
+            if (!rapsAsociados.isEmpty()) {
+                System.out.println("Entro aca");
+                throw new ReglaNegocioExcepcion("No se puede eliminar la competencia porque tiene RAPs asociados.");
+            }else{
+                competenciaRepositorio.deleteById(id);
+            }
             return true; // Eliminación exitosa
         } else {
-            throw new EntityNotFoundException("La competencia con el ID " + id + " no existe");
+            System.out.println("NO ENTRO");
+            throw new EntidadNoExisteException("La competencia con el ID " + id + " no existe");
         }
     }
+
+
     @Override
     public List<Competencia> obtenerCompetenciasPorIds(List<Integer> ids) {
         return competenciaRepositorio.findAllById(ids);
